+++ events/internal/controller/event/controller.go
@@ -3,10 +3,15 @@ package event
 import (
 	"context"
 	"errors"
+	"fmt"
+	"time"
 
-	eventmodel "pickfighter.com/events/pkg/model"
-	"pickfighter.com/pkg/pgxs"
 	"github.com/jackc/pgx/v5"
+	"github.com/spf13/viper"
+	"pickfighter.com/events/pkg/model"
+	eventmodel "pickfighter.com/events/pkg/model"
+	"pickfighter.com/events/pkg/version"
+	"pickfighter.com/pkg/pgxs"
 )
 
 // ErrNotFound is returned when a requested record is not found.
@@ -37,5 +42,20 @@ type Controller struct {
 func New(repo eventRepository) *Controller {
 	return &Controller{
 		repo: repo,
+	}
+}
+
+// HealthCheck returns the current health status of the application.
+// It includes information such as the app version, start time, uptime,
+// and a message indicating the application's health.
+func (c *Controller) HealthCheck() *model.HealthStatus {
+	return &model.HealthStatus{
+		AppDevVersion: version.DevVersion,
+		AppName:       version.Name,
+		Timestamp:     time.Now().Format(time.RFC1123),
+		AppRunDate:    version.RunDate,
+		AppTimeAlive:  time.Now().Unix() - version.RunDate,
+		Healthy:       true,
+		Message:       fmt.Sprintf("[%s]: I'm fine, thanks!", viper.GetString("app.name")),
 	}
 }