+++ pickfighter/internal/controller/pickfighter/utils.go
@@ -1,9 +1,14 @@
 package pickfighter
 
 import (
+	"fmt"
+	"time"
+
 	eventmodel "pickfighter.com/events/pkg/model"
-	gatewaymodel "pickfighter.com/pickfighter/pkg/model"
 	fightersmodel "pickfighter.com/fighters/pkg/model"
+	"pickfighter.com/pickfighter/pkg/model"
+	gatewaymodel "pickfighter.com/pickfighter/pkg/model"
+	"pickfighter.com/pickfighter/pkg/version"
 )
 
 func (c *Controller) getFightersIds(events []*eventmodel.Event) []int32 {
@@ -52,4 +57,43 @@ func (c *Controller) eventsPretify(events []*eventmodel.Event, fighters []*fight
 	}
 
 	return updatedEvents
+}
+
+func (c *Controller) GetAuthServiceHealthStatus() *model.HealthStatus {
+	status, err := c.authGateway.ServiceHealthCheck()
+	if err != nil {
+		return badHealthStatus("auth-service")
+	}
+
+	return status
+}
+
+func (c *Controller) GetEventServiceHealthStatus() *model.HealthStatus {
+	status, err := c.eventGateway.ServiceHealthCheck()
+	if err != nil {
+		return badHealthStatus("event-service")
+	}
+
+	return status
+}
+
+func (c *Controller) GetFightersServiceHealthStatus() *model.HealthStatus {
+	status, err := c.fightersGateway.ServiceHealthCheck()
+	if err != nil {
+		return badHealthStatus("fighters-service")
+	}
+
+	return status
+}
+
+func badHealthStatus(serviceName string) *model.HealthStatus {
+	return &model.HealthStatus{
+		AppDevVersion: version.DevVersion,
+		AppName:       serviceName,
+		Timestamp:     time.Now().Format(time.RFC1123),
+		AppRunDate:    0,
+		AppTimeAlive:  0,
+		Healthy:       false,
+		Message:       fmt.Sprintf("[%s]: I can't feel my legs!", serviceName),
+	}
 }