+++ pickfighter/pkg/model/mapper.go
@@ -3,7 +3,20 @@ package model
 import (
 	eventmodel "pickfighter.com/events/pkg/model"
 	fightersmodel "pickfighter.com/fighters/pkg/model"
+	"pickfighter.com/gen"
 )
+
+func HealthStatusFromProto(req *gen.HealthResponse) *HealthStatus {
+	return &HealthStatus{
+		AppDevVersion: req.AppDevVersion,
+		AppName:       req.AppName,
+		AppRunDate:    req.AppRunDate,
+		AppTimeAlive:  req.AppTimeAlive,
+		Healthy:       req.Healthy,
+		Message:       req.Message,
+		Timestamp:     req.Timestamp,
+	}
+}
 
 func ServiceEventToGatewayEvent(event *eventmodel.Event, fightersList map[int32]*fightersmodel.Fighter) *Event {
 	fights := event.Fights